requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alabaster==0.7.13
altair==5.5.0
annotated-types==0.7.0
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.1.0
bleach==6.0.0
blinker==1.9.0
bs4==0.0.1
cachelib==0.3.0
cachetools==5.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.0.1
cImage==2.0.2
click==8.0.1
comm==0.2.2
constantly==15.1.0
contourpy==1.0.7
cryptography==37.0.2
cssselect==1.2.0
curio==1.6
cycler==0.11.0
Cython==3.1.3
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
docrepr==0.2.0
docutils==0.18.1
docx2txt==0.9
durationpy==0.10
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.117.1
fastjsonschema==2.16.2
filelock==3.13.4
filetype==1.2.0
Flask==2.0.1
Flask-Session==0.4.0
flatbuffers==25.9.23
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.7.0
fsspec==2025.9.0
GeAI==0.1.8
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hyperlink==21.0.0
idna==3.3
imagesize==1.4.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib-resources==5.10.2
importlib_metadata==8.0.0
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.5
ipyparallel==8.4.1
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
libclang==18.1.1
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.7.0
matplotlib-inline==0.1.7
mistune==2.0.5
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.1.0
narwhals==1.47.1
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.6.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.3.1
ollama==0.6.1
openai==0.27.10
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
outcome==1.1.0
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.4
pathspec==0.11.0
patsy==1.0.1
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipenv==2023.12.1
platformdirs==4.2.2
plotly==6.3.1
pluggy==1.0.0
pmdarima==2.0.4
prometheus-client==0.16.0
prompt_toolkit==3.0.47
propcache==0.3.2
Protego==0.2.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==21.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pybase64==1.4.2
pycparser==2.21
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypdf==6.1.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==6.2.5
pytest-asyncio==0.20.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==2.0.6
pytz==2022.7.1
PyYAML==6.0
pyzmq==26.0.3
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
regex==2025.9.18
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
Scrapy==2.9.0
seaborn==0.13.2
selenium==4.1.5
Send2Trash==1.8.0
sentence-transformers==5.1.2
service-identity==21.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
streamlit==1.47.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.6.0
tiktoken==0.3.3
tinycss2==1.2.1
tk==0.1.0
tldextract==3.4.4
tokenizers==0.22.1
toml==0.10.2
tomli==2.0.1
torch==2.2.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.10.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uri-template==1.2.0
uritemplate==4.2.0
urllib3==1.26.9
uvicorn==0.36.0
uvloop==0.21.0
virtualenv==20.25.3
w3lib==2.1.1
wcwidth==0.2.13
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==15.0.1
Werkzeug==2.0.1
widgetsnbextension==4.0.5
wrapt==1.17.3
wsproto==1.1.0
xlrd==2.0.2
yarl==1.20.1
yfinance==0.2.22
zipp==3.19.2
zope.interface==6.0
zstandard==0.25.0
